+++ src/main/java/com/example/ecommerce/repository/orders/OrdersQueryImplementation.java
@@ -19,7 +19,7 @@ public class OrdersQueryImplementation implements OrdersQuery {
         cartItemsOfTheUser = jdbcTemplate.query("select cart_id,user_id,product_id from cart where user_id =" + userId, ((rs, rowNum) -> {
             Cart cart = new Cart();
             cart.setCartId(rs.getInt("cart_id"));
-            cart.setUserId(rs.getInt("user_id"));
+            cart.setUserId(rs.getLong("user_id"));
             cart.setProductId(rs.getInt("product_id"));
             return cart;
         }));
